Move current statedb to statedblegacy
