+++ vochain/state.go
@@ -17,8 +17,8 @@ import (
 	ed25519 "github.com/tendermint/tendermint/crypto/ed25519"
 	"go.vocdoni.io/dvote/crypto/ethereum"
 	"go.vocdoni.io/dvote/log"
-	"go.vocdoni.io/dvote/statedb"
-	"go.vocdoni.io/dvote/statedb/iavlstate"
+	statedb "go.vocdoni.io/dvote/statedblegacy"
+	"go.vocdoni.io/dvote/statedblegacy/iavlstate"
 	"go.vocdoni.io/dvote/types"
 	models "go.vocdoni.io/proto/build/go/models"
 	"google.golang.org/protobuf/proto"
